+++ src/NetTopologySuite.IO.GPX/GpxTrack.cs
@@ -39,6 +39,17 @@ namespace NetTopologySuite.IO
 
         public object Extensions { get; }
 
+        /// <inheritdoc />
+        public override string ToString() => Helpers.BuildString((nameof(this.Name), this.Name),
+                                                                 (nameof(this.Comment), this.Comment),
+                                                                 (nameof(this.Description), this.Description),
+                                                                 (nameof(this.Source), this.Source),
+                                                                 (nameof(this.Links), Helpers.ListToString(this.Links)),
+                                                                 (nameof(this.Number), this.Number),
+                                                                 (nameof(this.Classification), this.Classification),
+                                                                 (nameof(this.Segments), Helpers.BuildString((nameof(this.Segments.Length), this.Segments.Length))),
+                                                                 (nameof(this.Extensions), this.Extensions));
+
         internal static GpxTrack Load(XElement element, GpxReaderSettings settings)
         {
             if (element is null)